moved stuff around for clarity
